initial commit
This will not be the official repo for the project bonsai, but it
will help with tracking changes with the idea of bonsai. In the long
term I see bonsai becoming a set libraries and practices, in the short
term it seems best that it lives as an example project that grows.